refact: Trocando float para decimal

+++ Backend/Application/DTOs/OrderDTOs/OrderDTO.cs
@@ -5,8 +5,8 @@ namespace Application.DTOs.OrderDTOs
     public class OrderDTO
     {
         public Guid Id { get; set; }
-        public float SubTotal { get; set; }
-        public float Total { get; set; }
+        public decimal SubTotal { get; set; }
+        public decimal Total { get; set; }
         public IList<OrderItemDTO> Items { get; set; } = new List<OrderItemDTO>();
     }
 }
